+++ packages/sdk/src/chains/EVM/fee/index.ts
@@ -1,0 +1,3 @@
+export { calculateDynamicFee } from './dynamicFee';
+export { calculateBasicfee } from './basicFee';
+export { getFeeHandlerAddress } from './feeHandler';