+++ packages/sdk/src/types/types.ts
@@ -2,6 +2,7 @@ import { Bridge, ERC721MinterBurnerPauser } from '@buildwithsygma/sygma-contract
 import { ethers } from 'ethers';
 
 import { Erc20Detailed } from '../Contracts/Erc20Detailed';
+import { FeeType, TokenConfig } from '../chains/EVM';
 
 export interface SygmaSDK {}
 
@@ -13,29 +14,12 @@ export type Setup = {
   feeOracleSetup?: FeeOracleData;
 };
 
-export type TokenConfig = {
-  type: 'erc20' | 'erc721';
-  address: string;
-  name?: string;
-  symbol?: string;
-  imageUri?: string;
-  resourceId: string;
-  isNativeWrappedToken?: boolean;
-  decimals?: number;
-  isDoubleApproval?: boolean;
-  feeSettings: {
-    type: FeeType;
-    address: string;
-  };
-};
-
-export type FeeType = 'basic' | 'feeOracle' | 'none';
-
 export type EvmBridgeSetup = {
   name: string;
   type: 'Ethereum';
   networkId: number;
   bridgeAddress: string;
+  feeRouterAddress: string;
   erc20HandlerAddress: string;
   erc721HandlerAddress: string;
   feeOracleHandlerAddress?: string;
@@ -44,6 +28,7 @@ export type EvmBridgeSetup = {
   decimals: number;
   tokens: TokenConfig[];
   confirmations?: number;
+  feeHandlers: Array<{ address: string; type: 'basic' | 'oracle' }>;
 };
 
 export type EvmBridgeSetupList = EvmBridgeSetup[];